spring6mvc/src/main/java/dario/springframework/spring6restmvc/services/BeerCsvServiceImpl.java
package dario.springframework.spring6restmvc.services;

import com.opencsv.bean.CsvToBeanBuilder;
import dario.springframework.spring6restmvc.entities.Beer;
import dario.springframework.spring6restmvc.model.BeerCSVRecord;
import org.springframework.stereotype.Service;

import java.io.File;
import java.io.FileNotFoundException;
import java.io.FileReader;
import java.util.List;

@Service
public class BeerCsvServiceImpl implements BeerCsvService {
    @Override
    public List<BeerCSVRecord> convertCsv(File file) {
        try {
            return new CsvToBeanBuilder<BeerCSVRecord>(new FileReader(file))
                    .withType(BeerCSVRecord.class)
                    .build()
                    .parse();
        } catch (FileNotFoundException e) {
            throw new RuntimeException(e);
        }


    }
}
